Moved dates.js to components directory
